created the build-incides docker files and cleaned up the add-introns-to-gtf.py file
