+++ dockers/skylab/snss2-build-indices/modify_gtf_mouse.sh
