+++ dockers/skylab/snss2-build-indices/requirements.txt
@@ -1,0 +1,6 @@
+Cython==0.24.1
+pysam==0.16.0.1
+pytest-cov==2.10.1
+pytest==5.1.1
+black==19.3b0
+flake8==3.7.7
